refactor: Update CreateInvoiceWithProducts to use OmitType for products
The CreateInvoiceWithProducts class in the create-invoice-with-products.dto.ts file has been updated to use the OmitType utility from the @nestjs/swagger package. This change allows for the exclusion of the 'invoice_id' property from the CreateInvoiceProductsDto type when used as an array in the 'products' property of CreateInvoiceWithProducts. This improves the codebase by providing a more concise and expressive way to define the desired shape of the 'products' array.