refactor: Update InvoiceService to return InvoiceWithAllDataDto in get method
The InvoiceService in the invoice module has been updated to return an instance of InvoiceWithAllDataDto instead of Invoice in the get method. This change ensures that the complete invoice data, including all related data, is returned to the client. It improves the accuracy and completeness of the invoice retrieval process.