refactor: Add invoice preview route and service method
The InvoiceController has been updated to include a new route `invoicePreviewFromBody` which accepts a POST request and renders the `index` view. The `invoicePreviewFromBody` method in the InvoiceService has also been added to create an invoice preview based on the provided `createInvoiceDto`. This change enhances the functionality of the application by allowing users to preview invoices before submitting them.